+++ packages/driver/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- handle case with `spec.getCookies` throws an error when trying to get cookies of the browser
 
 ## 1.4.8 - 2022/1/5